removed DemoApplication.java and corresponding test. Created a new User Package and moved User entity and role to it
