+++ src/main/java/org/example/untitled/HomeController.java
@@ -1,4 +1,4 @@
-package org.example.untitled.controller;
+package org.example.untitled;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
