--- src/main/java/org/example/untitled/Mapper/UserMapper.java
@@ -1,5 +1,0 @@
-package org.example.untitled.Mapper;
-
-public class UserMapper {
-
-}
