+++ src/main/java/org/example/untitled/User/Mapper/UserMapper.java
@@ -1,0 +1,5 @@
+package org.example.untitled.User.Mapper;
+
+public class UserMapper {
+
+}
