+++ src/main/java/org/example/untitled/User/service/UserService.java
@@ -1,7 +1,7 @@
-package org.example.untitled.service;
+package org.example.untitled.User.service;
 
-import org.example.untitled.Mapper.UserMapper;
-import org.example.untitled.Repository.UserRepository;
+import org.example.untitled.User.Mapper.UserMapper;
+import org.example.untitled.User.Repository.UserRepository;
 import org.springframework.stereotype.Service;
 
 @Service
