+++ src/main/java/org/example/untitled/usercase/controller/CaseController.java
@@ -1,0 +1,4 @@
+package org.example.untitled.usercase.controller;
+
+public class CaseController {
+}
