+++ src/main/java/org/example/untitled/usercase/service/CaseService.java
@@ -1,0 +1,4 @@
+package org.example.untitled.usercase.service;
+
+public class CaseService {
+}
