--- src/test/java/org/example/untitled/DemoApplicationTests.java
@@ -1,13 +1,0 @@
-package org.example.untitled;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-//@SpringBootTest
-//class DemoApplicationTests {
-//
-//    @Test
-//    void contextLoads() {
-//    }
-//
-//}
