+++ src/test/java/org/example/untitled/UntitledApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class DemoApplicationTests {
+class UntitledApplicationTests {
 
     @Test
     void contextLoads() {
